Add files via upload
This works on all of my classes. I wasn't originally planning on having it do it for all of my classes, just one... so right now it's inefficient in the sense that it logs into BB for every class - as in, it opens up a new window and goes back to the login page and logs in again... for every class.